requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
pillow==11.1.0
setuptools==76.0.0
sqlparse==0.5.3
